refactor: relocate events logic to events class
