+++ lib/entity.js
@@ -1,6 +1,7 @@
 import { Interpreter, ClassInterpreter } from './generators/interpreter'
 import { Lexer } from './generators/lexer'
 import { escapeHTML } from './helpers/sanitize'
+import { Events } from './entity/events'
 
 export default class Entity {
   constructor(el) {
@@ -10,9 +11,10 @@ export default class Entity {
       className: el.className,
     }
     this.variables = []
-    this.listener = {}
     this.dynamicAttributes = []
     this.id = this.generateEntityUUID()
+
+    this.events = new Events(this)
 
     this._getDynamicAttributes()
 
@@ -203,7 +205,7 @@ export default class Entity {
 
   async init() {
     this.getVariables()
-    this.applyEventBindings()
+    this.events.apply()
     await this.evaluateAll()
 
     MiniJS.elements.push(this)
@@ -225,11 +227,6 @@ export default class Entity {
     }
   }
 
-  async evaluateEventAction(attrName) {
-    const attrVal = this.element.getAttribute(attrName)
-    await this._interpret(attrVal)
-  }
-
   async evaluateClass() {
     const expr = this.element.getAttribute(':class')
     if (!expr) return
@@ -238,12 +235,6 @@ export default class Entity {
       base: this.baseClasses,
       isClass: true,
     })
-  }
-
-  async evaluateLoadEvents() {
-    const loadExpr = this.element.getAttribute(':load')
-    if (!loadExpr) return
-    await this.evaluateEventAction(':load')
   }
 
   async evaluateEach() {
@@ -329,135 +320,8 @@ export default class Entity {
     }
   }
 
-  applyEventBindings() {
-    this.removeEventBindings()
-
-    const el = this.element
-
-    // Change binding
-    if (el.hasAttribute(':change')) {
-      this.listener[':change'] = {
-        el,
-        eventName:
-          el.type == 'checkbox' || el.tagName == 'select' ? 'change' : 'input',
-        event: () => {
-          this.evaluateEventAction(':change')
-        },
-      }
-    }
-
-    if (el.hasAttribute(':clickout')) {
-      this.listener[':clickout'] = {
-        el: document,
-        eventName: 'click',
-        event: (e) => {
-          if (!document.documentElement.contains(e.target)) return
-          if (el.contains(e.target)) return
-          this.evaluateEventAction(':clickout')
-        },
-      }
-    }
-
-    if (el.hasAttribute(':press')) {
-      this.listener[':press'] = []
-      this.listener[':press'].push({
-        el,
-        eventName: 'keyup',
-        event: (e) => {
-          if (e.target !== el) return
-          if (!['Enter', 'Space'].includes(e.code)) return
-          if (e.code == 'Space') e.preventDefault()
-          this.evaluateEventAction(':press')
-        },
-      })
-
-      this.listener[':press'].push({
-        el,
-        eventName: 'click',
-        event: (e) => {
-          this.evaluateEventAction(':press')
-        },
-      })
-
-      this.listener[':press'].push({
-        el,
-        eventName: 'touchstart',
-        event: (e) => {
-          this.evaluateEventAction(':press')
-        },
-      })
-    }
-
-    // Other Event Bindings
-    Array.from(el.attributes).forEach((attr) => {
-      if (
-        attr.name.startsWith(':') &&
-        !MiniJS.allCustomBindings.includes(attr.name)
-      ) {
-        const nativeEventName = attr.name.substring(1)
-        this.listener[attr.name] = {
-          el,
-          eventName: nativeEventName,
-          event: () => {
-            this.evaluateEventAction(attr.name)
-          },
-        }
-      } else if (attr.name.startsWith(':keyup.')) {
-        const [event, keycode] = attr.name.split('.')
-        const nativeEventName = event.substring(1)
-
-        let key = keycode[0].toUpperCase() + keycode.slice(1)
-
-        if (['up', 'down', 'left', 'right'].includes(keycode)) {
-          key = 'Arrow' + key
-        } else if (!['enter', 'space'].includes(keycode)) {
-          return
-        }
-
-        this.listener[attr.name] = {
-          el,
-          eventName: nativeEventName,
-          event: (e) => {
-            if (e.target !== el) return
-            if (e.key !== key) return
-            this.evaluateEventAction(attr.name)
-          },
-        }
-      }
-    })
-
-    // Add event listeners
-    Object.keys(this.listener).forEach((key) => {
-      const listener = this.listener[key]
-
-      if (Array.isArray(listener)) {
-        listener.forEach(({ el, eventName, event }) => {
-          el.addEventListener(eventName, event)
-        })
-      } else {
-        const { el, eventName, event } = listener
-        el.addEventListener(eventName, event)
-      }
-    })
-  }
-
   hasAttribute(attr) {
     return !!this.element.getAttribute(attr)
-  }
-
-  removeEventBindings() {
-    Object.keys(this.listener).forEach((key) => {
-      const listener = this.listener[key]
-
-      if (Array.isArray(listener)) {
-        listener.forEach(({ el, eventName, event }) => {
-          el.removeEventListener(eventName, event)
-        })
-      } else {
-        const { el, eventName, event } = listener
-        el.removeEventListener(eventName, event)
-      }
-    })
   }
 
   dispose() {
@@ -474,7 +338,7 @@ export default class Entity {
       if (!entity) continue
 
       variables.push(...entity.variables)
-      entity.removeEventBindings()
+      entity.events.dispose()
     }
 
     // Remove disposed elements
