+++ lib/main.js
@@ -172,7 +172,7 @@ const MiniJS = (() => {
 
   function _evaluateLoadEvents() {
     _elements.forEach((entity) => {
-      entity.evaluateLoadEvents()
+      entity.events.evaluate(':load')
     })
   }
 
@@ -192,7 +192,7 @@ const MiniJS = (() => {
 
   function _applyBindings() {
     _elements.forEach((entity) => {
-      entity.applyEventBindings()
+      entity.events.apply()
     })
   }
